+++ src/main/java/io/bootique/tapestry/TapestryModule.java
@@ -15,9 +15,9 @@ import io.bootique.jetty.JettyModule;
 import io.bootique.jetty.MappedFilter;
 import io.bootique.tapestry.annotation.IgnoredPaths;
 import io.bootique.tapestry.annotation.Symbols;
-import io.bootique.tapestry.annotation.TapestryFilter;
 import io.bootique.tapestry.annotation.TapestryModuleBinding;
 import io.bootique.tapestry.di.GuiceTapestryModule;
+import io.bootique.tapestry.filter.BQTapestryFilter;
 import io.bootique.tapestry.filter.BQTapestryFilterFactory;
 import org.apache.tapestry5.services.LibraryMapping;
 
@@ -28,6 +28,15 @@ import java.util.logging.Level;
 public class TapestryModule extends ConfigModule {
 
     /**
+     * @param binder DI binder passed to the Module that invokes this method.
+     * @return an instance of {@link TapestryModuleExtender} that can be used to load Tapestry custom extensions.
+     * @since 0.5
+     */
+    public static TapestryModuleExtender extend(Binder binder) {
+        return new TapestryModuleExtender(binder);
+    }
+
+    /**
      * Returns a MapBinder for Tapestry "symbols" that allow to reconfigure Tapestry runtime.
      *
      * @param binder DI binder to use for contributions.
@@ -35,11 +44,23 @@ public class TapestryModule extends ConfigModule {
      * @see org.apache.tapestry5.SymbolConstants
      * @see org.apache.tapestry5.internal.InternalConstants
      * @since 0.4
+     * @deprecated since 0.5 call {@link #extend(Binder)} and then call
+     * {@link TapestryModuleExtender#setSymbol(String, String)}.
      */
+    @Deprecated
     public static MapBinder<String, String> contributeSymbols(Binder binder) {
         return MapBinder.newMapBinder(binder, String.class, String.class, Symbols.class);
     }
 
+    /**
+     * Returns a MapBinder for Tapestry "symbols" that allow to reconfigure Tapestry runtime.
+     *
+     * @param binder DI binder to use for contributions.
+     * @return A new Multibinder.
+     * @deprecated since 0.5 call {@link #extend(Binder)} and then call
+     * {@link TapestryModuleExtender#addLibraryMapping(LibraryMapping)}.
+     */
+    @Deprecated
     public static Multibinder<LibraryMapping> contributeLibraries(Binder binder) {
         return Multibinder.newSetBinder(binder, LibraryMapping.class);
     }
@@ -51,7 +72,10 @@ public class TapestryModule extends ConfigModule {
      * @param binder DI binder to use for contributions.
      * @return a Multibinder for Tapestry modules.
      * @since 0.4
+     * @deprecated since 0.5 call {@link #extend(Binder)} and then call
+     * {@link TapestryModuleExtender#addTapestryModule(Class)}.
      */
+    @Deprecated
     public static Multibinder<Class<?>> contributeModules(Binder binder) {
         TypeLiteral<Class<?>> type = new TypeLiteral<Class<?>>() {
         };
@@ -62,33 +86,30 @@ public class TapestryModule extends ConfigModule {
      * @param binder DI binder to use for contributions.
      * @return Multibinder for URL paths that should not be processed by Tapestry.
      * @since 0.4
+     * @deprecated since 0.5 call {@link #extend(Binder)} and then call
+     * {@link TapestryModuleExtender#addIgnoredPath(String)}.
      */
+    @Deprecated
     public static Multibinder<String> contributeIgnoredPaths(Binder binder) {
         return Multibinder.newSetBinder(binder, String.class, IgnoredPaths.class);
     }
 
     @Override
     public void configure(Binder binder) {
-        JettyModule.contributeMappedFilters(binder).addBinding().to(Key.get(MappedFilter.class, TapestryFilter.class));
-
-        contributeLibraries(binder);
-        contributeIgnoredPaths(binder);
-        contributeSymbols(binder);
-        contributeModules(binder).addBinding().toInstance(GuiceTapestryModule.class);
+        TapestryModule.extend(binder).initAllExtensions().addTapestryModule(GuiceTapestryModule.class);
+        TypeLiteral<MappedFilter<BQTapestryFilter>> tf = new TypeLiteral<MappedFilter<BQTapestryFilter>>() {
+        };
+        JettyModule.extend(binder).addMappedFilter(tf);
 
         // decrease default verbosity...
-        BQCoreModule.contributeLogLevels(binder)
-                .addBinding("org.apache.tapestry5.modules.TapestryModule.ComponentClassResolver")
-                .toInstance(Level.WARNING);
-        BQCoreModule.contributeLogLevels(binder)
-                .addBinding("io.bootique.tapestry.filter.BQTapestryFilter")
-                .toInstance(Level.WARNING);
+        BQCoreModule.extend(binder)
+                .setLogLevel("org.apache.tapestry5.modules.TapestryModule.ComponentClassResolver", Level.WARNING)
+                .setLogLevel("io.bootique.tapestry.filter.BQTapestryFilter", Level.WARNING);
     }
 
     @Singleton
     @Provides
-    @TapestryFilter
-    MappedFilter createTapestryFilter(
+    MappedFilter<BQTapestryFilter> createTapestryFilter(
             ConfigurationFactory configurationFactory,
             Injector injector,
             @Symbols Map<String, String> diSymbols,